+++ packages/service-apollo/src/hooks.js
@@ -1,0 +1,5 @@
+const SERVICE_NAME = `apollo`;
+
+module.exports = {
+  SERVICE_NAME,
+};