+++ packages/service-apollo/src/lib/hooks.js
@@ -1,9 +1,0 @@
-const { SERVICE } = require("@forrestjs/hooks");
-
-const SERVICE_NAME = `${SERVICE} apollo`;
-// const FASTIFY_HEALTHZ_HANDLER = `${SERVICE_NAME}/hanlder`;
-
-module.exports = {
-  SERVICE_NAME
-  //   FASTIFY_HEALTHZ_HANDLER,
-};