+++ packages/service-fastify-apollo/src/lib/hooks.js
@@ -1,7 +1,0 @@
-const { SERVICE } = require("@forrestjs/hooks");
-
-const SERVICE_NAME = `${SERVICE} fastify-apollo`;
-
-module.exports = {
-  SERVICE_NAME
-};