+++ packages/service-hash/src/index.js
@@ -81,7 +81,7 @@ const serviceBcrypt = ({ registerAction }) => {
 
   // Fastify Integration (optional hook)
   registerAction({
-    hook: '$FASTIFY_HACKS_AFTER?',
+    hook: '$FASTIFY_HACKS_BEFORE?',
     name: hooks.SERVICE_NAME,
     trace: __filename,
     handler: ({ fastify }, { getContext }) => {