+++ packages/service-jwt/src/index.js
@@ -85,7 +85,7 @@ const serviceJwt = ({ registerAction }) => {
 
   // Fastify Integration (optional hook)
   registerAction({
-    hook: '$FASTIFY_HACKS_AFTER?',
+    hook: '$FASTIFY_HACKS_BEFORE?',
     name: hooks.SERVICE_NAME,
     trace: __filename,
     handler: ({ fastify }, { getContext }) => {