+++ packages/service-apollo/CHANGELOG.md
@@ -1,2 +1,5 @@
 # Change Log
 
+## v4.0.0
+
+- Refactored to remove babel and the build step